requirements.txt
CherryPy
jinja2
six
requests
ruamel.yaml
salt
ansible
ws4py